+++ BodilyFunctions_Workshop/Contents/mods/PZ_BodilyFunctions/42/media/lua/server/Camping/BF_Fuel.lua
@@ -1,2 +1,0 @@
-campingFuelType.HumanFeces = 0.2
-campingFuelType["HumanFeces"] = 0.2